<!--  Thanks for sending a pull request!  Here are some tips for you:
1. If this is your first time, please read our contributor guidelines: https://github.com/m3db/m3/blob/master/CONTRIBUTING.md and developer notes https://github.com/m3db/m3/blob/master/DEVELOPMENT.md
2. Please prefix the name of the pull request with the component you are updating in the format "[component] Change title" (for example "[dbnode] Support out of order writes") and also label this pull request according to what type of issue you are addressing. Furthermore, if this is a WIP or DRAFT PR, please create a draft PR instead: https://github.blog/2019-02-14-introducing-draft-pull-requests/
3. Ensure you have added or ran the appropriate tests for your PR: read more at https://github.com/m3db/m3/blob/master/DEVELOPMENT.md#testing-changes
4. Follow the instructions for writing a changelog note: read more at https://github.com/m3db/m3/blob/master/DEVELOPMENT.md#updating-the-changelog
-->

**What this PR does / why we need it**:
<!--
If you have an issue this change addresses, please add the following details:
*Automatically closes linked issue when PR is merged.
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->
Aggregator does not need an extremely accurate rate limiter, as it's only used to prevent overload.
Trading off some correctness* for speed - it can potentially block additional n*$numcpus requests during the window, which IMO is a fair tradeoff. It is also more GC-friendly, as struct itself is pointerless, and is only 16 bytes larger even with padding.

[*] almost all of the uses reset the limiter under write lock, so it's a not a problem.

With static time:
```
benchmark               old ns/op     new ns/op     delta
BenchmarkLimiter        3887          802           -79.37%
BenchmarkLimiter-2      9754          2110          -78.37%
BenchmarkLimiter-6      7510          2021          -73.09%
BenchmarkLimiter-12     5751          2049          -64.37%
BenchmarkLimiter-24     5766          2043          -64.57%
```

With actual calls to time.Now():
```
benchmark               old ns/op     new ns/op     delta
BenchmarkLimiter        11711         7677          -34.45%
BenchmarkLimiter-2      9414          5046          -46.40%
BenchmarkLimiter-6      8052          2344          -70.89%
BenchmarkLimiter-12     5443          1907          -64.96%
BenchmarkLimiter-24     5473          1912          -65.06%
```

**Special notes for your reviewer**:

**Does this PR introduce a user-facing and/or backwards incompatible change?**:
<!--
If no, just write "NONE" in the release-note block below.
If yes, a release note is required:
Enter your extended release note in the block below. If the PR requires additional action from users switching to the new release, include the string "action required".
-->
```release-note
NONE
```

**Does this PR require updating code package or user-facing documentation?**:
<!--
If no, just write "NONE" in the documentation-note block below.
If yes, describe which documentation you updated.
Note: Any changes that significantly updates how a code package is architectured or functions requires code package documentation updates in the form of updates to the README.md file in that package.
-->
```documentation-note
NONE
```